+++ userspace/include/net.h
@@ -12,7 +12,7 @@ typedef struct __attribute__((__packed__)){
     uint8_t     sender_hw_address[6];
     uint8_t     sender_proto_addr[4];
     uint8_t     target_hw_address[6];
-    uint16_t    target_proto_addr[4];
+    uint8_t     target_proto_addr[4];
 } ArpPacket;
 
 typedef struct __attribute__((__packed__)){
